ceph-windows: Added script for collecting Windows event logs
In order to collect more information for the build artifacts,
the collect-event-logs.ps1 script will be run in order to extract
the event logs from the Windows client machine.

It will dump all the event logs as evtx and then convert them to
txt in order to be accessible on all platforms.

If the -IncludeEvtxFiles flag is used, the evtx files can be kept.
By default they are deleted.

Also, if the -CleanupEventLog flag is used, then the machine's
events will get cleared after the dump. By default they are kept.

The dumped event log files can be found in the directory sent as
parameter using the mandatory -LogDirectory parameter.

Signed-off-by: Stefan Chivu <schivu@cloudbasesolutions.com>