ceph-windows: Build Ceph on Windows inside libvirt VM
Make sure that the Ceph on Windows build is done inside an Ubuntu
libvirt VM. This ensures that the build is done on a clean environment
on each job run.

After the Windows build is done, the Ubuntu VM is rebuilt to ensure
that Ceph vstart will have a clean environment as well.

Signed-off-by: Ionut Balutoiu <ibalutoiu@cloudbasesolutions.com>