increase vm memory for windows test jobs
Windows test jobs use two vms: a windows vm that takes 8gb of ram
and a linux one that currently uses 32gb of ram.

We're using memstore with 5gb per osd. It seems that the linux vm
is running out of memory, which is why the OSDs can get
terminated while running the tests.

We'll go ahead and increase the linux vm memory to 64gb.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>