scripts/build_utils.sh: pin urllib3 and requests
The latest version of chacractl requires requests-toolbelt==0.9.1.
This is not compatible with the latest version of urllib3, which
dropped support for appengine.

We need to pin urllib3 to an older version that still supports
appengine, and requests to an older version that still supports
the older urllib3 until charactl allows a higher version of
requests-toolbelt.

Fixes: https://tracker.ceph.com/issues/59652
Signed-off-by: Laura Flores <lflores@redhat.com>