ceph-windows-pull-requests: constrain to windows label
Currently, the intersection of focal && libvirt labels returns some
smithi (nproc == 8) and mira (nproc == 4) nodes, which are inadequate
for the two VMs that the Windows job spawns.  It turns we already have
a windows label, which, after dropping it from mira015, seems to be on
point (i.e. would filter those low CPU nodes out).  Let's use it!

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>