Keep track of extra script get-tarballs.sh
This is used by sync-pull to fetch the tarballs when
constructing a release, but needs modification for the
prerelease/ directory structure.  It's invoked by ssh
to download.ceph.com to run from /home/signer/bin, because
it needs to copy large files from chacra to download.ceph.com
and doing that on download.ceph.com is much more efficient.

Signed-off-by: Dan Mick <dan.mick@redhat.com>