Add ceph-dev-pipeline-trigger
This job serves to trigger ceph-dev-pipeline via webhook payloads, but is
opt-in. This implementation initially targets only GitHub push event payloads.
On each push, it will inspect the head commit for [git trailers][0]. The trailer
required to opt in is "CI-PIPELINE: true". If that trailer is found, the job
looks for these others:

ARCHS, CI_COMPILE, CI_CONTAINER, CI_PIPELINE, DISTROS, DWZ, FLAVOR, SCCACHE

Any that are found are passed as parameters.

[0] https://git-scm.com/docs/git-interpret-trailers

Signed-off-by: Zack Cerza <zack@cerza.org>