Update all jobs with ${VAR} syntax to ${{VAR}} for jjb6 compatibility
jenkins-job-builder 6 changed the interpretation of {var} to *always*
refer to a jjb macro; this is true even for ${VAR}, which we use in a
lot of place for Jenkins variable expansion (although not all; it's
unclear whether the {} are or ever were necessary, or in what exact
contexts).  So far, we've pinned jjb to <6, but we need new
functionality.

In any event, rather than changing all of them to the same {}-less style
that might work, change the ${} style to ${{}} and leave $ alone.

Signed-off-by: Dan Mick <dan.mick@redhat.com>