ansible: Fixes to podman subuid/subgid checks
I think this was working during my testing because sub{u,g}id_check.rc /wasn't/ 0.  I had manually modified the system so it wasn't in the same state as a fresh cloud instance.

I manually ran this playbook on an instance and confirmed it gets past the podman setup now.

Signed-off-by: David Galloway <david.galloway@ibm.com>