ceph-release-container: add setup_container_runtime.sh
A ceph-release-container run suffered from

error running container: from /usr/bin/crun creating container for
[<sh command>]: sd-bus call: Interactive authentication required.:
Permission denied

That is thought to be because of the lack of 'linger' on the
jenkins-build user, such that the /run/user directory didn't
persist.  The exact mechanism of failure isn't fully understood,
but this script has fixed similar issues for ceph-dev-pipeline.

Also, add an echo of the node for convenience.

Signed-off-by: Dan Mick <dan.mick@redhat.com>