ceph-tag: Write a sha1 file for later consumption
In a regular build, ceph-dev-pipeline just takes a branch name and makes a source tarball in ceph-source-dist.  ceph-dev-pipeline takes that tarball from the ceph-source-dist job and builds from it so it's able to just get the SHA1 from it.

In a /release/ build, a release branch (e.g., reef-release) is specified and we add an additional version commit on top of it.  So we don't know the SHA1 just based on the branch name.  My solution was to write the version commit's SHA1 to a file and consume it later.

Signed-off-by: David Galloway <david.galloway@ibm.com>