builder.yml: update to use virtualenv
pip/python no longer like to install modules in global dirs.
Create a virtualenv and use it when running the jenkins_node
module.

This is done by adding a second play to start that creates the
virtualenv and sets ansible_python_interpreter to the interpreter
in the venv.  That means it affects the following play in the
playbook, which is the original playbook.  So technically the
virtualenv python is used for all plays.

Also, remove the #!/usr/bin/python3 from library/jenkins_node,
which was doing no good, but caused havoc by execing the non-venv
python when running the module.

Signed-off-by: Dan Mick <dan.mick@redhat.com>