sync-pull: Check for all distro/ver subdirs and create if necessary
The previous logic was fine for the first time a release was sync-pulled as long as that new release doesn't have a new RPM distro/distrover.

e.g., New tentacle release, old logic saw there was no tentacle-21.0.0 dir and created debian/jessie and centos/9 subdirs.  But we later needed to add rocky/10 and create a rocky/10 subdir.  But the logic saw tentacle-21.0.0 so didn't create it.

Signed-off-by: David Galloway <david.galloway@ibm.com>