Update ceph-website-prs.yml to comment with website link on success
it seems like this was commented out to reduce noisy comments on website site pr's. 

Right now the workflow is to go look at the jenkins job console output and find the link. So having a update comment saying the site has been successfully built and can be view at xxx is a better workflow.