ceph-dev-pipeline: check shaman for "are there binaries"
Checking chacra isn't the right check, since this build's
chacra url is not necessarily the prior successful build's
chacra url (there are multiple chacra instances).  Instead,
check shaman to see if there is a built repo; if so, the build
can be skipped.  build_container will search shaman again to
find the repo url to use to build the container, if necessary.

Signed-off-by: Dan Mick <dan.mick@redhat.com>