ceph-dev-pipeline: Mark several vars as Fields
When variables are defined using 'def' at the top level of a Jenkinsfile, they
are added to the scope local to the generated pipeline function. When they
are defined as Fields, they are instead marked as instance variables of the
encompassing class. As a result, they do not increase the size of the function,
avoiding the MethodTooLargeException. A tradeoff is that their values reset if a
pipeline run is restarted midway, but that's not a feature we are using.

Signed-off-by: Zack Cerza <zack@cerza.org>