Introduce "debug" flavor
Currently, the only flavor used for testing is "default".
While this ensures that the tested flavor matches the released flavor, it can also be a limitation.
Introducing a debug flavor would allow us to test branches that require additional or more thorough validation.
The main difference is that built-in assertions would be compiled in.

The reasons for this change are:
a) The Crimson suite uses a crimson-debug flavor for project PR gating.
   The next commit will clean up the Crimson flavor entirely,
   and the new debug flavor introduced here could be used as its replacement.

b) Good practice: having an additional build with debug enabled is useful
   when retesting or performing extra checks.

Initially, the new debug flavor will only apply to centos9 builds.
If it proves valuable, we can expand support to other distros.

Note: The current way to schedule debug builds is by using a *-debug branch name.
      Having a dedicated flavor seems more straightforward.

Signed-off-by: Matan Breizman <mbreizma@redhat.com>