ceph-dev-new-setup/build: Introduce crimson-tentacle builds
See the following comment:
```
Tentacle is the last release that needs dedicated Crimson builds,
Later releases are able to use Crimson with the default build.
As the "Crimson flavor" is no longer available, we need a *temporary* way
to be able build Crimson for tentacle.
Note: This could be removed once Crimson we have Umbrella release builds.
```

Signed-off-by: Matan Breizman <mbreizma@redhat.com>