ceph-windows: Update qcow2 image location. Bump VM spec.
The previous one was living on the LRC and served via reverse proxy.  The RDU lab was shut down earlier and I neglected to back the qcow2 image up.  Who knows when the old LRC will be brought back online so the image has been rebuilt.

Also bumping the VM resources.

Signed-off-by: David Galloway <david.galloway@ibm.com>